feat(positions): multichain positions (#5093)
### Description

gets positions and shortcuts for whichever networks are enabled in our
dynamic config, rather than just celo

relies on https://github.com/valora-inc/hooks/pull/362

### Test plan

- ci
- manual test with local hooks api on wallet with both ethereum and celo
positions (also unearthed a redux migration bug with this! hooray for
testing!)

### Related issues

fixes
https://linear.app/valora/issue/RET-941/%5Bhooks%5D-multichain-support

### Backwards compatibility

there is only one place where the new API will affect application code
for old app versions, which is in the `keyExtractor` function in
`AssetList.tsx`. The function returns the following for positions:
`${activeTab}-${item.appId}-${item.networkId}-${item.address}-${index}`.
I think this is ok, since old app versions only get celo positions right
now, so when `item.networkId` evaluates to `undefined` for them, it
shouldn't cause collisions with other positions on other networks.

the other changes are all in analytics events, which should be something
we can live with

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)