feat(points): Add initial points activity feed saga (#5206)
### Description

For
[RET-1032](https://linear.app/valora/issue/RET-1032/[wallet-fake-backend]-points-activity-history).
Adds in saga/data fetching for points history. (I also started
implementing the bottom sheet itself but left it out here in the
interest of size, see videos below. Will get out a PR after this as a
fast follow.)

See the related PR for the backend `/getHistory` endpoint
[here](https://github.com/valora-inc/points-functions/pull/23).

### Test plan

Unit and manual tested. See videos below.

#### Happy path (no error)


https://github.com/valora-inc/wallet/assets/569401/3fd18685-8f27-4b09-b1b9-227eb1fc92de

#### Error


https://github.com/valora-inc/wallet/assets/569401/72838322-f454-4df0-b49d-e838f984d0d2

### Related issues

- Part of
[RET-1032](https://linear.app/valora/issue/RET-1032/[wallet-fake-backend]-points-activity-history).

### Backwards compatibility

Yes.
### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)

---------

Co-authored-by: Kathy Luo <kathyluo18@gmail.com>