fix(qr-scanner): Allow data with network prefixes (#5262)
### Description

just ignores the network prefix and reads the address

I considered trying to read and validate the network, but some things
stopped me:
1. lack of any kind of plumbing for a 'default network' in the send flow
to action upon the prefix. If we want to use it, should probably be
another ticket and have specific UX guidelines
2. Lack of consistency for how other wallets mark their data.

| Wallet              | Celo | Ethereum |
| :---------------- | :------: | ----: |
| Metamask        |   ethereum:0x   | ethereum:0x |
| Coinbase           |  celo:0x   | ethereum:0x |
| SafeWallet    |  celo:0x   | ethereum:0x |

Sometimes the wallets mark L2 chains as their own (celo, optimism,
arbitrum) and sometimes just label it 'ethereum'. This inconsistency
makes it difficult to action upon.

### Test plan

Unit tested with examples pulled from metamask and coinbase

### Related issues

- Fixes
https://linear.app/valora/issue/ACT-1155/chain-prefix-support-for-qr-scanner

---------

Co-authored-by: Satish Ravi <satish.ravi@valoraapp.com>