ci: cancel outdated merge queue requests, run test always on main (#5275)
### Description

Fixes a couple of things
1. Cancels in progress merge queue requests on outdated branches. We
originally had this because we used only `head_ref` as the key, which is
defined only for PR runs. But now we also have `ref_name`, which should
be unique per merge queue PR
2. Run Test workflow on main. Sometimes we merge skipping the queue
because of codecov/patch failures (or other reasons). In this case,
subsequent PRs won't have a base to compare against because codecov
never ran on the commit on main, leading to inconsistent results.

### Test plan

CI

### Related issues

N/A

### Backwards compatibility

N/A

### Network scalability

N/A