feat(points): Add points history bottom sheet (#5254)
### Description

Part of
[RET-1032](https://linear.app/valora/issue/RET-1032/[wallet-fake-backend]-points-activity-history).
Adds in a points history bottom sheet, minus the actual card rendering.

### Test plan

Unit and manual tested. See videos:


https://github.com/valora-inc/wallet/assets/569401/7319dce2-934a-4bf2-95db-4bb583136430



https://github.com/valora-inc/wallet/assets/569401/1766cddb-50d2-43da-bb10-54e5df6a7b4a



### Related issues

Part of
[RET-1032](https://linear.app/valora/issue/RET-1032/[wallet-fake-backend]-points-activity-history).

### Backwards compatibility

Yes.

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)