feat(wallet-tab): include error message when token fetching fails (#5308)
### Description

The home token balance component included a toast error message when
token fetching fails. This adds the same to the wallet tab. Also added
more tests for AssetsTokenBalance (used in the wallet tab) and cleaned
up some redundant tests in preparation for removing the home token
balance (part of cleaning up the drawer nav)

### Test plan

Unit tests, manually failing the token fetch request

### Related issues

- Part of ACT-1133

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)