fix(feed): use tx status from the blockchain-api (#5323)
### Description

Use transaction status returned from the blockchain-api instead of the
hardcoded `TransactionStatus.Success` value.

Does not invalidate local transactions cache expecting it to be updated
as soon as there is an internet connection.

### Test plan

* Tested manually


![image](https://github.com/valora-inc/wallet/assets/2737872/e84669d3-bda6-4853-8563-c3c9be93588f)

### Related issues

* Fixes RET-1057
* Fixes ACT-1167

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)