feat(points): Add error banner on failure fetching subsequent points history pages (#5362)
### Description

Adds in an inline notification error when there's a failure fetching
subsequent points history page.s See designs
[here](https://www.figma.com/file/rXBDplfMHHqYmuu6EkgMEo/Gamification-experiments?type=design&node-id=1486-6438&mode=design&t=gVtyIzOM0LmxJaDy-4).

### Test plan

Unit and manual tested. See video below.


https://github.com/valora-inc/wallet/assets/569401/c5cb9fc4-50a3-424c-84f6-dab6fb10f981

### Related issues

- Fixes #[issue number here]

### Backwards compatibility

Yes.

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)