feat(swap): enable app fee via remote config (#5359)
### Description

This PR allows enabling the swap fee via the remote config.

When enabled, the free Valora swap fee line item is hidden.
The details of the Valora fee are then displayed when hitting the info
button of the exchange rate line item.

### Test plan

- Updated tests

| Valora fee disabled | Valora fee enabled (0 fee from API) | Valora fee
enabled (positive fee from API) |
|----------|----------|----------|
|
![image](https://github.com/valora-inc/wallet/assets/57791/6cded654-86bd-46eb-ab8d-da35076588bd)
|
![image](https://github.com/valora-inc/wallet/assets/57791/038fb3b4-d2cf-4d49-8325-40b6824038a8)
|
![image](https://github.com/valora-inc/wallet/assets/57791/576fcf93-b99a-4269-a387-fa009e10dedb)
|
|
![image](https://github.com/valora-inc/wallet/assets/57791/37ac4758-4d25-46a7-852a-c6b21ea44047)
|
![image](https://github.com/valora-inc/wallet/assets/57791/49cd7a09-775e-4b47-91b4-ecf0a022b4c6)
|
![image](https://github.com/valora-inc/wallet/assets/57791/dd66d5d3-0bd3-4c89-a5cd-4b4b1c34d442)
|

Notes: 
- copies are not final
- the "Valora fee disabled" case is same as today except it also shows
the info icon next to the exchange rate.

### Related issues

- Fixes RET-1061

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)