fix(points): prevent sagas spawning if points are disabled (#5385)
### Description

As the title. Rather than remembering to check the remote config in each
saga, catch it all at the root level. We should not start talking to the
points DB yet.

### Test plan

n/a

### Related issues

n/a

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)