feat: earn asset details for aaveUSDC (#5403)
### Description

Expands on the work in #5400 to allow the multiple button types and
pools in `EarnCard.tsx`. This abstraction should now work for any Aave
v3 pools on any network and sets up the ability to handle multiple pools
with only layout changes needed when designs are ready.

### Test plan

- [x] Tested locally on iOS
- [x] Tested locally on Android
- [x] Unit tests added and updated

### Related issues

- Fixes ACT-1185

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)