feat(earn): Add enter amount screen (#5399)
### Description

Add EarnEnterAmount screen.

### Test plan
Unit tests added.

**Manual tests:**

When a user has USDC on arbitrum:


https://github.com/valora-inc/wallet/assets/140328381/5103c167-3447-4df5-bb86-2318bf2c021a



When a user doesn't have enough USDC on arbitrum:


https://github.com/valora-inc/wallet/assets/140328381/459a6671-e8bc-42ca-84ef-e8c6d104a427



Clicking the info button:


https://github.com/valora-inc/wallet/assets/140328381/86c47948-61c2-4c1b-89b9-b02be943f534



### Related issues

- Fixes #ACT-1176

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [X] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)