feat(swap): Initial Swap screen redesign (#5409)
### Description

First part of
[RET-1065](https://linear.app/valora/issue/RET-1065/[wallet]-swap-screen-redesign).
Updates the design of swap screen to match the [designs in
Figma](https://www.figma.com/design/fjpfHZdVSLazv330kT2sQz/Cross-chain-swaps?node-id=606-1653&t=Bq4e7bCLkjpGu4D0-4).
The rest of the behavior outlined in the ticket will come in followup
PRs.

### Test plan

Unit and manual tested. See video below:



https://github.com/valora-inc/wallet/assets/569401/c078eb34-4cd2-4fe9-a9a2-f5eba85c0c47



### Related issues

- Part of
[RET-1065](https://linear.app/valora/issue/RET-1065/[wallet]-swap-screen-redesign).

### Backwards compatibility

Yes.

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)