chore(points): add track points event to transaction watcher (#5441)
### Description

In case the transaction is not settled within the same app session, we
should add the points tracking to the transaction watcher as well.

I added the network to the tracked event, in preparation for network
specific activities.

### Test plan

n/a

### Related issues

n/a

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)