fix(points): ensure balance is displayed timely and correctly (#5459)
### Description

This PR fixes:
1. the points balance was outdated on the discover tab because the
balance fetch was only triggered on the points home screen. now we also
trigger a balance fetch on app mount.
2. also fetch points config on app mount (consolidate logic, was
previously being fetched on home screen visit)
3. disable font scaling on the points balance because the font size is
large already, and font scaling messes with the animation.

### Test plan

n/a

### Related issues

n/a

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)