fix(points): fetch points balance and history on backup restore (#5604)
### Description

Start fetching points balance and history on backup restore. Otherwise
users will not see their points balance when they navigate to the
Dicover section right after restoring their wallet.

Context:
https://valora-app.slack.com/archives/C04B61SJ6DS/p1719509875907719?thread_ts=1719507689.657979&cid=C04B61SJ6DS

| Before | After |
|--------|--------|
|
![](https://github.com/valora-inc/wallet/assets/2737872/fa53ba3a-8962-4b6d-9881-93d755c960ed)
|
![](https://github.com/valora-inc/wallet/assets/2737872/c77d25f4-0e1b-4bb5-8b23-f3e206fcf065)
|



### Test plan

* Tested manually

### Related issues

NA

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)