chore!: remove DAppKit (#5622)
### Description

It's time to say good bye to DAppKit!
We thank you for being the very first way to interact with DApps on
Celo.

- DAppKit has been sunset since march 2023 
  - https://forum.celo.org/t/sunsetting-dappkit/5336
- no dapps on the dapps list are actively using it

### Test plan

- Tests pass

### Related issues

- Fixes RET-853
- Fixes ACT-1101
- Fixes ACT-1088

### Backwards compatibility

No, but nobody's using DAppKit anymore.

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)