feat(points): Add support for earning Points for sending Live Links (#5605)
### Description

For
[RET-1079](https://linear.app/valora/issue/RET-1079/enable-points-for-live-links-send).
Adds support in-app for earning points on Live Link sends, and rendering
history.

Note that both the backend and the Jumpstart contract support ERC721
sends, though it's currently not possible in-app to use the Jumpstart
system to send NFTs. Despite this, since it's supported everywhere else,
ERC721 Live Link history is supported in-app.

### Test plan

Unit and manual tested, see video below. Note that the reclaim in the
video is meant to demonstrate that points are _not_ granted for anything
other than sends, even though sends/claims both interact with the
Jumpstart contract.



https://github.com/valora-inc/wallet/assets/569401/1477002f-a1dc-4b84-8300-51a959e73240


### Related issues

- Fixes
[RET-1079](https://linear.app/valora/issue/RET-1079/enable-points-for-live-links-send).

### Backwards compatibility

Yes.

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)