feat(swap): update swap info UI and add transaction time for cross chain (#5626)
### Description

This PR implements part of [this
design](https://www.figma.com/design/fjpfHZdVSLazv330kT2sQz/Cross-chain-swaps?node-id=606-2259&t=7bMhTAB8SD1gWv6z-0).
The fees part of this design is not included in this PR.
- updated UI on the swaps screen
- updated bottom sheet styles
- added estimated transaction time for cross chain quotes

### Test plan


https://github.com/valora-inc/wallet/assets/20150449/76c7ca36-ee66-4744-bfd6-d3c6b1787d1c




### Related issues

- Related to RET-1066

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)