feat(onboarding): Recovery phrase screen update (#5642)
### Description

Update copy and styles to match designs.

Checked with design and confirmed that we should not have the "Why is
this important?" line, and rather just keep the old help button + bottom
sheet.

### Test plan

![Simulator Screenshot - iPhone 14 Pro - 2024-07-12 at 14 46
45](https://github.com/user-attachments/assets/8ca705c6-9200-4355-8510-8e384c8f2b02)


### Related issues

- Fixes #ACT-1221

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [X] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)

---------

Co-authored-by: Tom McGuire <Mcgtom10@gmail.com>