feat(points): Reduce trackPointsEvent usage (#5643)
### Description

Taking another crack at #5631. Updated following a
[discussion](https://github.com/valora-inc/wallet/pull/5631#pullrequestreview-2168324821)
with @kathaypacific. Now, events that are already emitted simply have
properties required by the point system added on, without encoding any
points semantics explicitly.

### Test plan

Unit and manual tested; points are still earned for performing expected
tasks.

### Related issues

- Fixes #[issue number here]

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)