fix(hooks): too many claim shortcuts displayed (#5658)
### Description

With the switch of hooks to multichain, shortcuts needed to be filtered
by `networkId`.
Also the displayed tokens needed an update to use `tokenId`.

### Test plan

| Before | After |
|-|-|
|
![image](https://github.com/user-attachments/assets/c54378f3-c015-45e3-97c0-2fb3f300b58a)
|
![image](https://github.com/user-attachments/assets/c1c09a0f-add3-466a-a2d9-24c1ab5f9825)
|

### Related issues

- Fixes RET-1108

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)