chore: update token bottom sheet chain select to auto dismiss (#5644)
### Description

This PR makes a few minor tweaks / fixes to the bottom sheet network
filter:
1. if only a single selection is allowed, dismiss the bottom sheet on
selection - this is more of a preference thing, but it seems unnecessary
(and maybe unexpected) to require the user to explicitly dismiss the
bottom sheet when only one option can be selected. The auto dismissing
behaviour is what we are used to on iOS (not sure about Android, but
seems nice to save a tap regardless)
2. the "no results" message should display the network name (I think
that is more helpful than the current message of `no results found for
"Network"`)
3. fix the missing unique react keys warning
4. make the analytics more accurate by dispatching the
`network_filter_updated` event on network selection. currently the
`network_filter_updated` event is sent on the `onClose` of the
BottomSheet. so a) the event is dispatched twice after selecting a
network and pressing "done", b) the event is dispatched even if the user
does not make a selection (e.g. they just open the filter and dismiss
it)
5. use `getBy*` for positive test assertions

I verified that these changes don't impact the version of the component
that allows multiple selections (aside from the analytics change)

### Test plan



https://github.com/user-attachments/assets/81756f27-4c59-4a0e-84b4-3d7a86512f97



### Related issues

n/a

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)