feat: allow `tokensByIdSelector` to return position tokens and count tokens once for total balance calculation (#5659)
### Description

This contains the following changes:

- Update `tokensByIdSelector`:
- When a regular token has no `priceUsd`, we update it with the value
sourced from position tokens if we have it
- This allows showing $ price for complex ERC20 tokens when the position
can be decomposed via position hooks.
- Include tokens from positions when `includePositionTokens: true` is
passed
- `useTokenInfo` and `getTokenInfo` do that now. This shouldn't affect
existing behaviors though.
    - This allows `TokenDisplay` to display tokens from positions
- Tokens/positions are only counted once when calculating total
balances:
  - The total tokens balance ignores tokens sourced from positions
- The total positions balance ignores tokens present in the regular list

Note: tokens from positions still can't be seen in the "Tokens" tab, or
sent or swapped, unless explicitly added to address-metadata (or
manually imported). We can decide later whether or not we want to enable
that.

More generally, existing token selectors still return the same data
(except with $ price coming from position if we have it), and only
places that call `tokensByIdSelector` passing `includePositionTokens:
true` get position tokens too.

This was ported from https://github.com/valora-inc/troopo/pull/17

### Test plan

- Updated tests

### Related issues

- Fixes RET-1113
- Fixes RET-1114

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)