refactor(statsig): pull multichain features into its own function to decrease risk of crashes (#5669)
### Description

* Ensure at compile that developers cannot use `getDynamicConfigParams`
to get multichain dynamic config
* Add in a new `getMultichainFeatures` function to retrieve the dynamic
config for multichain and filter out unknown networkids

Almost all of the file changes are just updating
`getDynamicConfigParams` to `getMultichainFeatures`

### Test plan

unit tests updated.

### Related issues


https://linear.app/valora/issue/ACT-1078/decrease-risk-of-wallet-crash-when-adding-new-networks

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)