feat(earn): Add EarnHome screen (#5656)
### Description

Move over necessary multi-pool earn stuff from earnV2. Two pools shown
for testing purposes but the PR only includes the single Aave pool
currently supported.

Will add analytics and hook stuff up in follow-up PRs

### Test plan


https://github.com/user-attachments/assets/c6a76424-367f-4d26-84ec-7e0343dd50c0

| Open pools | My pools | Bottom sheet |
| ----- | ----- | ----- | 
|
![open](https://github.com/user-attachments/assets/4e9b6d0a-9bfb-40e3-bf96-44341ef116af)
|
![my](https://github.com/user-attachments/assets/d5ff397a-5f80-4d40-9523-5e5194cb48da)
|
![bottom-sheet](https://github.com/user-attachments/assets/6d739afc-0885-4eff-9ecc-273bdfeaf9aa)
|

### Related issues

- Part of ACT-1260

### Backwards compatibility

Yes
### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [X] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)

---------

Co-authored-by: Satish Ravi <satish.ravi@valoraapp.com>