refactor(statsig): remove defaults for feature gates (#5689)
### Description

Unlike experiments and dynamic config, the feature gate sdk call
`checkGate` doesn't take default params as an option. If the SDK is
uninitialized it will always return `false`, so we should always model
gates to default to false. The fallback default we currently have is
only used in case of exceptions, which is a very unlikely scenario. This
PR removes the default values and updates the checkGate exception
handler to return false for all gates (except the 2 that currently
defaults to true)

### Test plan

CI

### Related issues

- N/A

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)