refactor: make FeeRowItem reusable for all fee types (#5693)
### Description

This PR should have no user facing changes. This change makes it easier
to add the fees on the cross chain swap tx detils.

### Test plan

n/a

### Related issues

- Related to RET-1119

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)