feat(earn): fetch earn positions + selector (#5679)
### Description

Fetch and store earn positions. Also add a selector to list them.

User and earn positions are fetched and stored together.

For each returned earn position:
- if the user already holds the earn position, it is ignored (since it's
already present as a user position)
- if not it's added to the list

Next up: use earn positions to drive the earn screen (which was added in
https://github.com/valora-inc/wallet/pull/5656)

### Test plan

- Updated tests

### Related issues

- Fixes RET-1121

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)