feat(shortcuts): include positionId when triggering shortcuts (#5700)
### Description

This change could be useful in the future to properly identify the
position that triggered the shortcut.

For instance we could imagine the hooks backend could retrieve the full
position from the id, to pass it to the `onTrigger` method of the
shortcut. Giving additional context that could be needed.

### Test plan

- Updated tests

### Related issues

- Fixes RET-1110

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)