refactor: use useAllowedNetworks in a hook (#5703)
### Description

This was a suggestion in #5699 but it was merged before being addressed.
This also fixes a bad trailing space that was left in the string, which
causes a 400 response from blockchain-api.

### Test plan

n/a

### Related issues

n/a

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)