chore: update audit known issues (ws) (#5707)
### Description

Update the known issues. The diff is a bit hard to review but there are
no _new_ ignored vulnerabilities here, the references have just changed.
From what I can see, the nature of the reported vulnerabilities have
not. Also see #5333.

### Test plan

CI passes

### Related issues

n/a

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)