fix(swap): ensure long running cross chain swaps eventually shows completed status (#5704)
### Description

This PR fixes a bug where pending cross chains swaps remain pending
forever. The root cause is the query helper reconciles already received
transactions with newly received transactions by preferring the already
received transaction. This means that previously received pending swaps
are not updated when the same transactions is received again with the
completed status.

This is a bug that affects not only cross chain swaps (we should always
prefer the newest information in case other fields in the transaction
have changed.

### Test plan

Cross chains swaps can be settled correctly

### Related issues

n/a

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)