fix: display default network explorer in tx details for failed cross chain swap (#5708)
### Description

The only time we mark cross chain swaps as failed is when it fails on
the source network. In this case, the cross chain swap hasn't started
and AxelarScan does not have a reference to it. We should use the
default network explorer which has more information on why the tx
failed.

### Test plan

![Simulator Screenshot - iPhone 15 Pro - 2024-08-05 at 13 14
48](https://github.com/user-attachments/assets/02eca7bf-222f-4eb6-9a08-dcb201839717)


### Related issues

n/a

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)