fix: handle received pending cross chain swaps correctly (#5716)
### Description

This PR fixes a couple of bugs:
1. when a cross chain swap is pending for a long time, the app vibrates
periodically for no apparent reason
2. when a user restores their wallet and there is a pending cross chain
swap, and it is actually completed, it will show completed in the app
session but on subsequent app sessions, it always starts off as pending.

The root cause for both is due to the query helper attributing the cross
chain swap as new or existing incorrectly. For the scenarios:
1. the pending cross chain swap isn't stored with the other received
blockchain-api transactions (it is instead stored with the pending
standby transactions) so every poll interval, we receive the pending tx
from blockchain-api -> it's not in the redux store that we expect -> we
count it as new and vibrate
2. the pending tx received from blockchain-api doesn't exist in the
standby redux store -> we store it in redux with the other received
blockchain-api tx's -> later we receive from blockchain-api the updated
status as complete -> we never update the cached transactions because
this transaction [doesn't meet the definition of new
transaction](https://github.com/valora-inc/wallet/blob/6ce4465733f676fb2ec01acbabe7f2d7d9ca81c9/src/transactions/feed/queryHelper.ts#L196).

This PR fixes both by including a check for the transaction status when
determining whether or not to dispatch the redux action. This is the
second time I've had to fix something around the queryHelper so I also
tried to add some tests focused on the action dispatch and haptic
feedback.

### Test plan

n/a

### Related issues

n/a

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)