feat(earn): Use hooks info for multi-pool earn flow (#5730)
### Description

Use hooks earn positions in the multi-pool earn flow rather than
hard-coded pool list.

Updated logic so it uses balance from the hooks response rather than
tokenInfo, that way we don't need to add new pool tokens to
address-metadata.

Deposit and withdraw flows still need to be updated to work for
multi-pool.

### Test plan

Tests updated 


https://github.com/user-attachments/assets/bd50e7fa-6e03-416f-9326-5f70c286c586



### Related issues

- Fixes ACT-1268

### Backwards compatibility

Yes, updating code behind feature gate not yet rolled out

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [X] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)