fix: update CYA options and styles (#5768)
### Description

Changes: 
- Replace old links with the new ones as per task requirement.
- Style updates

Notes:
- Created `LogoHeartConfigurable` with configurable colour as a
duplicate of `LogoHeart` as it is added to `.gitignore`.
- Links were already shuffled by `shuffle` function in code so the order
will differ from the one in Figma.

### Test plan
<img
src="https://github.com/user-attachments/assets/82dc56c0-0c37-475c-a2b6-1dc686c6a669"
width="300" />


### Related issues

- Fixes RET-1175

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)